Add first version of a rudimentary test suite
The two source files are used to check if the compiled onnxruntime can be used
and produce the correct results.
The Interference_cnn should return 10 values which shoulkd sum up to a
probability of one. These information si alreday checked at the end of the
function and the return values are choosen accordingly.
This test can be used to check if the onnxruntime installation works as
expected.
The result of the second test is currently not understood and I have asked
for more information from the developer of the two tests.